requirements.txt
py-clob-client>=0.34
python-dotenv
requests
